# api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: Communication with Slack API
  version: '1'
  title: Slack API
host: localhost:10010
basePath: /
schemes:
  - https
paths:
...
  /updateMsg:
    x-swagger-router-controller: updateMsg
    post:
      operationId: updateMsg
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              ts:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /updateMsg/oneButton:
    x-swagger-router-controller: updateMsg
    post:
      operationId: updateMsgOneButton
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              ts:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /updateMsg/oneButton/confirm:
    x-swagger-router-controller: updateMsg
    post:
      operationId: updateMsgOneButtonConfirm
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textConfirmation1:
                type: string
              ts:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /updateMsg/twoButtons:
    x-swagger-router-controller: updateMsg
    post:
      operationId: updateMsgTwoButtons
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textButton2:
                type: string
              ts:
                type: string

      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
  /updateMsg/twoButtons/Confirm:
    x-swagger-router-controller: updateMsg
    post:
      operationId: updateMsgTwoButtonsConfirm
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            properties:
              channel:
                type: string
              text:
                type: string
              callbackId:
                type: string 
              textButton1:
                type: string
              textButton2:
                type: string
              textConfirmation1:
                type: string
              textConfirmation2:
                type: string
              ts:
                type: string
      responses:
        '200':
          description: Definition generated from Swagger Inspector
          schema:
            type: string
...